requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
sqlalchemy-utils
flake8
pyjwt
passlib[bcrypt]